requirements.txt
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
billiard==3.3.0.23
celery==3.1.26.post2
cffi==1.11.5
cryptography==2.3.1
Django==2.1.1
django-celery==3.2.2
django-cors-headers==2.4.0
django-redis==4.9.0
djangorestframework==3.8.2
idna==2.7
kombu==3.0.37
Pillow==5.2.0
pycparser==2.18
PyMySQL==0.9.2
pytz==2018.5
redis==2.10.6
six==1.11.0
